conduit: Implement `RequestExt` for `Request<Cursor<Bytes>>`
This allows us to get rid of the `ConduitRequest` struct, since we can use `Request<Cursor<Bytes>>` directly
